fix(auth) method setCookie to service
