+++ src/auth/auth.controller.ts
@@ -13,10 +13,8 @@ import {
 } from '@nestjs/common';
 import { AuthService } from './auth.service';
 import { CreateAuthDto } from './dto/create-auth.dto';
-import { UpdateAuthDto } from './dto/update-auth.dto';
-import { Response, CookieOptions } from 'express';
+import { Response } from 'express';
 import { IJwtConfig, jwtConfig } from './../config/jwt.config';
-import ms from 'ms';
 
 @Controller('auth')
 export class AuthController {
@@ -33,27 +31,7 @@ export class AuthController {
     @Res({ passthrough: true }) res: Response,
   ) {
     const { user, tokens } = await this.authService.register(createAuthDto);
-    this.setAuthCookies(res, tokens);
+    this.authService.setAuthCookies(res, tokens);
     return { user };
-  }
-
-  private setAuthCookies(
-    res: Response,
-    tokens: { accessToken: string; refreshToken: string },
-  ) {
-    const cookieOptions: CookieOptions = {
-      httpOnly: true,
-      secure: false,
-      sameSite: 'lax',
-    };
-
-    res.cookie('accessToken', tokens.accessToken, {
-      ...cookieOptions,
-      maxAge: ms(this.jwtConfigService.accessExpiresIn),
-    });
-    res.cookie('refreshToken', tokens.refreshToken, {
-      ...cookieOptions,
-      maxAge: ms(this.jwtConfigService.refreshExpiresIn),
-    });
   }
 }
